move threaded code to new workers/ module
